Added a command line parser, added instructions on how to use the chat client in the README, made the syntax more consisten
