+++ Makefile
@@ -1,4 +1,13 @@
-all:
-	gcc -g -Wall -Werror -pedantic-errors chatclient.c -o chatclient
+CC=gcc
+CFLAGS=-Wall -Werror -pedantic-errors
+
+all: chatclient chatserver
+
+chatclient: src/chatclient.c
+	$(CC) $(CFLAGS) src/chatclient.c -o chatclient
+
+chatserver: src/chatserver.c
+	$(CC) $(CFLAGS) src/chatserver.c -o chatserver
+
 clean:
-	rm -f chatclient chatclient.exe
+	rm -f chatclient chatserver
